Move shared instance message checks to separate class MODINVSTOR-1001
