+++ src/test/java/org/folio/rest/api/InstanceStorageTest.java
@@ -18,12 +18,12 @@ import static org.folio.rest.support.http.InterfaceUrls.instancesStorageUrl;
 import static org.folio.rest.support.http.InterfaceUrls.natureOfContentTermsUrl;
 import static org.folio.rest.support.matchers.DateTimeMatchers.hasIsoFormat;
 import static org.folio.rest.support.matchers.DateTimeMatchers.withinSecondsBeforeNow;
-import static org.folio.rest.support.matchers.DomainEventAssertions.deleteAllEventForInstancesPublished;
-import static org.folio.rest.support.matchers.DomainEventAssertions.instanceCreatedMessagePublished;
-import static org.folio.rest.support.matchers.DomainEventAssertions.instanceCreatedMessagesPublished;
-import static org.folio.rest.support.matchers.DomainEventAssertions.instanceDeletedMessagePublished;
-import static org.folio.rest.support.matchers.DomainEventAssertions.instancedUpdatedMessagePublished;
-import static org.folio.rest.support.matchers.DomainEventAssertions.noInstanceMessagesPublished;
+import static org.folio.rest.support.messages.InstanceEventMessageChecks.instanceCreatedMessagesPublished;
+import static org.folio.rest.support.messages.InstanceEventMessageChecks.noInstanceMessagesPublished;
+import static org.folio.rest.support.messages.InstanceEventMessageChecks.deleteAllEventForInstancesPublished;
+import static org.folio.rest.support.messages.InstanceEventMessageChecks.instancedUpdatedMessagePublished;
+import static org.folio.rest.support.messages.InstanceEventMessageChecks.instanceCreatedMessagePublished;
+import static org.folio.rest.support.messages.InstanceEventMessageChecks.instanceDeletedMessagePublished;
 import static org.folio.rest.support.matchers.PostgresErrorMessageMatchers.isMaximumSequenceValueError;
 import static org.folio.rest.support.matchers.PostgresErrorMessageMatchers.isUniqueViolation;
 import static org.folio.util.StringUtil.urlEncode;
@@ -93,6 +93,7 @@ import org.folio.rest.support.ResponseHandler;
 import org.folio.rest.support.builders.HoldingRequestBuilder;
 import org.folio.rest.support.builders.ItemRequestBuilder;
 import org.folio.rest.support.db.OptimisticLocking;
+import org.folio.rest.support.messages.InstanceEventMessageChecks;
 import org.folio.rest.tools.utils.OptimisticLockingUtil;
 import org.folio.utility.LocationUtility;
 import org.junit.After;
